Moved file_utils outside of ml_worker module
