+++ giskard/datasets/base/__init__.py
@@ -1,5 +1,3 @@
-from typing import Dict, Hashable, List, Optional, Union
-
 import inspect
 import logging
 import posixpath
@@ -14,6 +12,7 @@ import pandas as pd
 import yaml
 from mlflow import MlflowClient
 from pandas.api.types import is_list_like, is_numeric_dtype
+from typing import Dict, Hashable, List, Optional, Union
 from xxhash import xxh3_128_hexdigest
 from zstandard import ZstdDecompressor
 
@@ -29,9 +28,8 @@ from giskard.registry.transformation_function import (
     TransformationFunctionType,
 )
 from giskard.settings import settings
-
-from ...ml_worker.utils.file_utils import get_file_name
 from ..metadata.indexing import ColumnMetadataMixin
+from ...utils.file_utils import get_file_name
 
 try:
     import wandb  # noqa
