+++ tests/utils.py
@@ -19,10 +19,10 @@ from giskard.client.giskard_client import GiskardClient
 from giskard.core.savable import Artifact
 from giskard.datasets.base import Dataset
 from giskard.ml_worker import ml_worker
-from giskard.ml_worker.utils.file_utils import get_file_name
 from giskard.models.base.model import BaseModel
 from giskard.path_utils import get_size
 from giskard.settings import settings
+from giskard.utils.file_utils import get_file_name
 
 logger = logging.getLogger(__name__)
 resource_dir: Path = Path.home() / ".giskard"
